Moved example code
