+++ integration-example/src/main/java/org/mache/examples/GettingCacheClient.java
@@ -1,8 +1,7 @@
-package org.mache;
+package org.mache.examples;
 
-import org.mache.examples.cassandra.CassandraAnnotatedMessage;
+import org.mache.ExCache;
 import org.mache.examples.cassandra.CassandraExample;
-import org.mache.examples.mongo.MongoAnnotatedMessage;
 import org.mache.examples.mongo.MongoExample;
 
 import java.util.Arrays;
@@ -10,38 +9,35 @@ import java.util.Arrays;
 /**
  * Created by jbowkett on 17/07/15.
  */
-public class PuttingCacheClient {
+public class GettingCacheClient {
+
   private enum CacheType {Cassandra, Mongo}
 
   public static void main(String...commandLine) {
     final Args args = parseArgs(commandLine);
     final int count = args.count;
+    final Example example;
     switch(args.cacheType){
       case Cassandra:
-        populateWithCassandraMsgs(count, new CassandraExample().exampleCache());
+        example = new CassandraExample();
         break;
       case Mongo:
-        populateWithMongoMsgs(count, new MongoExample().exampleCache());
+        example = new MongoExample();
         break;
       default:
         throw new RuntimeException("Invalid cache type: ["+args.cacheType+"].  Valid values are:"+Arrays.toString(CacheType.values()));
     }
+    doExample(count, example);
   }
 
-  private static void populateWithMongoMsgs(int count, ExCache<String, MongoAnnotatedMessage> cache) {
-    System.out.println("Putting...");
-    for(int i = 0; i< count ; i++){
-      final MongoAnnotatedMessage v = new MongoAnnotatedMessage("msg_" + i, "Hello World - " + i);
-      cache.put(v.getPrimaryKey(), v);
-    }
-  }
-
-  private static <T> void populateWithCassandraMsgs(int count, ExCache<String, CassandraAnnotatedMessage> cache)  {
-    System.out.println("Putting...");
+  private static <T> void doExample(int count, Example<T> example) {
+    final ExCache<String, T> cache = example.exampleCache();
+    System.out.println("Getting...");
     for (int i = 0; i < count ; i++) {
-      final CassandraAnnotatedMessage v = new CassandraAnnotatedMessage("msg_" + i, "Hello World - " + i);
-      cache.put(v.getPrimaryKey(), v);
+      final T hello = cache.get("msg_"+i);
+      System.out.println("hello = " + hello);
     }
+    cache.close();
   }
 
   private static Args parseArgs(String[] args) {
@@ -51,7 +47,7 @@ public class PuttingCacheClient {
       return new Args(count, cacheType);
     }
     else{
-      throw new RuntimeException("Usage : PuttingCacheClient <put count> "+ Arrays.toString(CacheType.values()));
+      throw new RuntimeException("Usage : GettingCacheClient <get count> "+ Arrays.toString(CacheType.values()));
     }
   }
